Extract Optional to its own header file
